bug fixes + more fleshed out implementation of the wave bell
wave bell should work correctly + more fleshed out implementation:

sypnosis of how the bell works:
Each encounter will have a set of waves, each with a certain amount of enemies

A wave will have a counter value and enemy value. The counter value is the amount of time before the wave bell will summon a wave of enemies decided by the enemy value.

Once a wave depletes, it will attempt to summon the next wave, if it cant (field is full), wave counter will be locked at 0 and will only move to the next wave once it is able to.

using a high cost card higher than the current wave's counter will not cause the counter to overflow to the next wave. This is intended as it incentivizes usage of high cost cards right before a wave ends.

various bug fixes:
Buttons now should not fire multiple inputs when spam clicking

proper delay when summoning enemy cards at the start of battle