created a "pause" menu
created a menu that is more of a hamburger menu than a pause menu, as there are issues with pausing the game at the moment. This menu is very simple, allowing you to exit back to the main menu or close it.

More polish would be added to it in the future. Specifically changes to the UI box would be made.